refactor(cli): extract wrapper, zccache, and binaries modules (#339)
Brings main.rs under the 1000-line target by lifting the RUSTC_WRAPPER
invocation path, the zccache build-session orchestration / subprocess
helpers, and the toolchain/rustup binary resolution helpers into
dedicated modules.

Refs #339
