+++ crates/soldr-cli/src/main.rs
@@ -1,7 +1,8 @@
 use clap::{Parser, Subcommand};
-use soldr_core::{suppress_windows_console_window, SoldrError, SoldrPaths};
+use soldr_core::{suppress_windows_console_window, SoldrError};
 use soldr_fetch::VersionSpec;
 
+mod binaries;
 mod cache;
 mod cargo_front_door;
 mod doctor;
@@ -10,11 +11,20 @@ mod linker;
 mod rust_plan;
 mod self_relocate;
 mod toolchain;
+mod wrapper;
+mod zccache;
 
-const TEST_CARGO_BIN_ENV_VAR: &str = "SOLDR_TEST_CARGO_BIN";
-const TEST_RUSTC_BIN_ENV_VAR: &str = "SOLDR_TEST_RUSTC_BIN";
-const TEST_RUSTUP_BIN_ENV_VAR: &str = "SOLDR_TEST_RUSTUP_BIN";
-const TEST_ZCCACHE_BIN_ENV_VAR: &str = "SOLDR_TEST_ZCCACHE_BIN";
+#[allow(unused_imports)]
+pub(crate) use binaries::{
+    apply_implicit_toolchain_homes, cached_managed_zccache, current_soldr_binary,
+    fetch_managed_zccache, non_empty_env_path, parse_tool_spec, resolve_toolchain_binary,
+    rustup_binary, rustup_resolution_failure, zccache_binary_override,
+};
+
+pub(crate) const TEST_CARGO_BIN_ENV_VAR: &str = "SOLDR_TEST_CARGO_BIN";
+pub(crate) const TEST_RUSTC_BIN_ENV_VAR: &str = "SOLDR_TEST_RUSTC_BIN";
+pub(crate) const TEST_RUSTUP_BIN_ENV_VAR: &str = "SOLDR_TEST_RUSTUP_BIN";
+pub(crate) const TEST_ZCCACHE_BIN_ENV_VAR: &str = "SOLDR_TEST_ZCCACHE_BIN";
 pub(crate) const TEST_FREE_DISK_BYTES_ENV_VAR: &str = "SOLDR_TEST_FREE_DISK_BYTES";
 /// Overrides the default `nightly` toolchain used by `soldr gc cargo`
 /// to invoke cargo's unstable `-Zgc clean gc`. The `--toolchain` flag
@@ -22,14 +32,14 @@ pub(crate) const TEST_FREE_DISK_BYTES_ENV_VAR: &str = "SOLDR_TEST_FREE_DISK_BYTE
 pub(crate) const SOLDR_GC_CARGO_TOOLCHAIN_ENV_VAR: &str = "SOLDR_GC_CARGO_TOOLCHAIN";
 pub(crate) const JSON_SCHEMA_VERSION: u32 = 1;
 pub(crate) const LOW_DISK_WARNING_THRESHOLD_BYTES: u64 = 2 * 1024 * 1024 * 1024;
-const RUSTC_WRAPPER_OVERRIDE_ENV_VAR: &str = "SOLDR_RUSTC_WRAPPER";
+pub(crate) const RUSTC_WRAPPER_OVERRIDE_ENV_VAR: &str = "SOLDR_RUSTC_WRAPPER";
 pub(crate) const CARGO_PROFILE_DEV_DEBUG_ENV_VAR: &str = "CARGO_PROFILE_DEV_DEBUG";
 pub(crate) const CARGO_PROFILE_TEST_DEBUG_ENV_VAR: &str = "CARGO_PROFILE_TEST_DEBUG";
 /// Picks the linker injected for `soldr cargo ...` builds. See `linker` module
 /// and `docs/API.md` for the supported values (`default | ld | mold | rust-lld
 /// | fast`).
 pub(crate) const LINKER_ENV_VAR: &str = "SOLDR_LINKER";
-const REAL_TOOLCHAIN_BINARY_ENV_PREFIX: &str = "SOLDR_REAL_";
+pub(crate) const REAL_TOOLCHAIN_BINARY_ENV_PREFIX: &str = "SOLDR_REAL_";
 pub(crate) const TARGET_CACHE_MODE_ENV_VAR: &str = "SOLDR_TARGET_CACHE_MODE";
 pub(crate) const TARGET_CACHE_BUNDLE_DIR_ENV_VAR: &str = "SOLDR_TARGET_CACHE_BUNDLE_DIR";
 pub(crate) const TARGET_CACHE_BACKEND_ENV_VAR: &str = "SOLDR_TARGET_CACHE_BACKEND";
@@ -397,7 +407,7 @@ async fn main() {
         }
     }
 
-    if raw_args.len() > 1 && is_wrapper_invocation(&raw_args[1]) {
+    if raw_args.len() > 1 && wrapper::is_wrapper_invocation(&raw_args[1]) {
         if let Some(version) = soldr_as_env_pin() {
             if should_trampoline(&version) {
                 std::process::exit(
@@ -407,7 +417,7 @@ async fn main() {
                 );
             }
         }
-        std::process::exit(run_rustc_wrapper(&raw_args).unwrap_or_else(report_and_exit));
+        std::process::exit(wrapper::run_rustc_wrapper(&raw_args).unwrap_or_else(report_and_exit));
     }
 
     // `--as <version>` trampoline. Peeled off before clap so the fetched
@@ -476,7 +486,10 @@ async fn run_cli(cli: Cli) -> Result<(), SoldrError> {
             std::process::exit(toolchain::run_toolchain_passthrough("rustfmt", &args)?);
         }
         Commands::ClippyDriver { args } => {
-            std::process::exit(toolchain::run_toolchain_passthrough("clippy-driver", &args)?);
+            std::process::exit(toolchain::run_toolchain_passthrough(
+                "clippy-driver",
+                &args,
+            )?);
         }
         Commands::Rustdoc { args } => {
             std::process::exit(toolchain::run_toolchain_passthrough("rustdoc", &args)?);
@@ -488,7 +501,10 @@ async fn run_cli(cli: Cli) -> Result<(), SoldrError> {
             std::process::exit(toolchain::run_toolchain_passthrough("rust-lldb", &args)?);
         }
         Commands::RustAnalyzer { args } => {
-            std::process::exit(toolchain::run_toolchain_passthrough("rust-analyzer", &args)?);
+            std::process::exit(toolchain::run_toolchain_passthrough(
+                "rust-analyzer",
+                &args,
+            )?);
         }
         Commands::Rustup { args } => {
             std::process::exit(toolchain::run_rustup_passthrough(&args)?);
@@ -536,11 +552,7 @@ async fn run_cli(cli: Cli) -> Result<(), SoldrError> {
                 // Either flag pair maps onto the same boolean; `dry_run`
                 // is the documented default so we accept it explicitly.
                 let _ = dry_run;
-                cache::run_cache_prune_target_command(
-                    path,
-                    effective_dry_run,
-                    prune_json || json,
-                )?;
+                cache::run_cache_prune_target_command(path, effective_dry_run, prune_json || json)?;
             }
             None => {
                 let output = cache::collect_cache_output()?;
@@ -669,7 +681,10 @@ fn should_self_relocate_for_invocation(raw_args: &[String]) -> bool {
             "cargo" => {
                 return cache_enabled
                     && cargo_front_door::cargo_args_are_cacheable(&args[idx + 1..])
-                    && matches!(rustc_wrapper_mode(), RustcWrapperMode::ManagedZccache);
+                    && matches!(
+                        zccache::rustc_wrapper_mode(),
+                        zccache::RustcWrapperMode::ManagedZccache
+                    );
             }
             _ => return false,
         }
@@ -800,1043 +815,6 @@ async fn run_trampoline(version: &str, args: &[String]) -> Result<i32, SoldrErro
 
         Ok(status.code().unwrap_or(1))
     }
-}
-
-/// Known toolchain binaries that cargo may invoke through RUSTC_WRAPPER
-/// or RUSTC_WORKSPACE_WRAPPER. When soldr is set as a wrapper, cargo
-/// passes: `soldr <toolchain-binary> <rustc-args...>`
-const WRAPPER_PASSTHROUGH_TOOLS: &[&str] = &["rustc", "clippy-driver"];
-
-fn is_wrapper_invocation(arg: &str) -> bool {
-    let stem = std::path::Path::new(arg)
-        .file_stem()
-        .and_then(std::ffi::OsStr::to_str)
-        .unwrap_or(arg);
-
-    WRAPPER_PASSTHROUGH_TOOLS.contains(&stem)
-}
-
-fn run_rustc_wrapper(raw_args: &[String]) -> Result<i32, SoldrError> {
-    let tool_arg = raw_args
-        .get(1)
-        .ok_or_else(|| SoldrError::Other("missing tool path in wrapper mode".into()))?;
-
-    let tool_stem = std::path::Path::new(tool_arg.as_str())
-        .file_stem()
-        .and_then(std::ffi::OsStr::to_str)
-        .unwrap_or(tool_arg);
-
-    // Per-build target/ tracking for `soldr gc`. Best-effort: if we
-    // can't resolve a workspace target dir cheaply, or the redb
-    // upsert fails for any reason, skip silently — never fail a build.
-    if tool_stem == "rustc" {
-        record_target_dir_in_registry(&raw_args[2..]);
-    }
-
-    // When the source argument is "-" (stdin), rustc reads the source from
-    // the process's stdin. If we pass this invocation to zccache as-is,
-    // zccache reads stdin to hash the source content, exhausting the pipe
-    // before rustc is spawned. Rustc then receives an empty stdin, compiles
-    // nothing, and exits 0 — masking any real compile error (e.g. E0554 from
-    // build-script feature probes like rustix 0.37's `can_compile()`).
-    //
-    // Fix: spill stdin to a temp file so both zccache and rustc see a real
-    // path. The temp file is created in the system temp directory and removed
-    // after the child exits. This keeps zccache in the loop (it can hash the
-    // file normally) while preserving the correct exit code.
-    let stdin_tempfile = if raw_args[2..].iter().any(|a| a == "-") {
-        Some(spill_stdin_to_tempfile()?)
-    } else {
-        None
-    };
-
-    // Build the effective arg list, replacing "-" with the temp file path.
-    let effective_args: std::borrow::Cow<[String]> = if let Some(ref tmp) = stdin_tempfile {
-        let tmp_str = tmp.path().to_string_lossy().into_owned();
-        let replaced: Vec<String> = raw_args
-            .iter()
-            .cloned()
-            .map(|a| if a == "-" { tmp_str.clone() } else { a })
-            .collect();
-        std::borrow::Cow::Owned(replaced)
-    } else {
-        std::borrow::Cow::Borrowed(raw_args)
-    };
-
-    // Only route through zccache for actual rustc invocations, not
-    // clippy-driver or other workspace wrappers.
-    if tool_stem == "rustc" && soldr_cache::cache_enabled_in_current_process() {
-        if let Some(zccache) = zccache_binary_override() {
-            return run_wrapper_through_zccache(&effective_args, &zccache);
-        }
-    }
-
-    // Resolve the tool binary. If it's already a full path, use it
-    // directly. Otherwise resolve via rustup.
-    let tool_path: std::path::PathBuf = if std::path::Path::new(tool_arg.as_str()).is_absolute() {
-        tool_arg.into()
-    } else {
-        resolve_toolchain_binary(tool_stem)?
-    };
-
-    let mut command = std::process::Command::new(tool_path);
-    command.args(&effective_args[2..]);
-    apply_implicit_toolchain_homes(&mut command);
-    suppress_windows_console_window(&mut command);
-    let status = command.status()?;
-
-    Ok(status.code().unwrap_or(1))
-}
-
-/// Read all of stdin into a named temporary file and return the file.
-///
-/// The file has a `.rs` extension so rustc accepts it without flags, and
-/// lives in the system temp directory. It is deleted when the returned
-/// `NamedTempFile` value is dropped (i.e. after the child process exits).
-fn spill_stdin_to_tempfile() -> Result<tempfile::NamedTempFile, SoldrError> {
-    use std::io::{Read, Write as _};
-    let mut tmp = tempfile::Builder::new()
-        .prefix("soldr-stdin-")
-        .suffix(".rs")
-        .tempfile()
-        .map_err(|e| SoldrError::Other(format!("failed to create stdin temp file: {e}")))?;
-    let mut buf = Vec::new();
-    std::io::stdin()
-        .read_to_end(&mut buf)
-        .map_err(|e| SoldrError::Other(format!("failed to read stdin: {e}")))?;
-    tmp.write_all(&buf)
-        .map_err(|e| SoldrError::Other(format!("failed to write stdin temp file: {e}")))?;
-    Ok(tmp)
-}
-
-
-
-fn run_wrapper_through_zccache(
-    raw_args: &[String],
-    zccache: &std::path::Path,
-) -> Result<i32, SoldrError> {
-    let mut command = std::process::Command::new(zccache);
-    command.args(&raw_args[1..]);
-    suppress_windows_console_window(&mut command);
-
-    // Cargo's jobserver lives on numbered file descriptors that it inherits
-    // into the RUSTC_WRAPPER, advertised via CARGO_MAKEFLAGS. On Unix,
-    // exec'ing into zccache replaces the wrapper process in-place so those
-    // FDs flow straight through to the inner rustc — rustc otherwise emits
-    // "failed to connect to jobserver from environment variable
-    // CARGO_MAKEFLAGS=...: cannot open file descriptor N" because spawning
-    // a Rust child closes any FDs not explicitly inherited.
-    #[cfg(unix)]
-    {
-        use std::os::unix::process::CommandExt;
-        // TODO(#265): once exec() replaces this process there is nowhere to
-        // observe zccache's stderr or retry on "unknown session:". The
-        // Windows branch below performs that defensive retry. A Unix port
-        // would need to spawn-with-piped-stderr instead of exec, while still
-        // forwarding the cargo jobserver FDs (see issue #265 for context).
-        let err = command.exec();
-        Err(SoldrError::Other(format!(
-            "failed to exec zccache at {}: {err}",
-            zccache.display()
-        )))
-    }
-
-    #[cfg(not(unix))]
-    {
-        run_wrapper_through_zccache_windows(raw_args, zccache)
-    }
-}
-
-/// Windows-only wrapper invocation: spawn zccache with its stderr piped so we
-/// can tee it to our own stderr live AND scan it after the process exits.
-///
-/// If zccache returns a non-zero exit and its stderr contains the literal
-/// substring `unknown session:` (issue #265), the managed zccache daemon was
-/// killed mid-build by something outside soldr's control (e.g. zccache-ci's
-/// stop hook on older zccache, AV quarantine, or a Windows binary
-/// replacement). We allocate a fresh session via `zccache session-start` and
-/// retry the wrapper invocation exactly once with the new session id.
-///
-/// Retry budget is 1. On the retry's own failure we propagate that exit code
-/// unchanged — we don't loop on a persistently broken daemon.
-#[cfg(not(unix))]
-fn run_wrapper_through_zccache_windows(
-    raw_args: &[String],
-    zccache: &std::path::Path,
-) -> Result<i32, SoldrError> {
-    use std::io::Read;
-    use std::process::Stdio;
-
-    let mut command = std::process::Command::new(zccache);
-    command.args(&raw_args[1..]);
-    command.stderr(Stdio::piped());
-    suppress_windows_console_window(&mut command);
-
-    let mut child = command.spawn()?;
-    let stderr = child
-        .stderr
-        .take()
-        .expect("stderr was configured as piped above");
-
-    // Tee zccache stderr to soldr's stderr in real time AND buffer it for
-    // post-exit inspection. A reader thread keeps the pipe drained so
-    // zccache cannot block on a full pipe.
-    let reader = std::thread::spawn(move || -> std::io::Result<Vec<u8>> {
-        let mut buf = Vec::new();
-        let mut reader = std::io::BufReader::new(stderr);
-        let mut chunk = [0u8; 4096];
-        loop {
-            let n = reader.read(&mut chunk)?;
-            if n == 0 {
-                break;
-            }
-            // Best-effort tee: if writing to our own stderr fails we still
-            // want to keep draining the child pipe.
-            let _ = std::io::Write::write_all(&mut std::io::stderr(), &chunk[..n]);
-            buf.extend_from_slice(&chunk[..n]);
-        }
-        Ok(buf)
-    });
-
-    let status = child.wait()?;
-    let stderr_bytes = reader
-        .join()
-        .map_err(|_| SoldrError::Other("zccache stderr reader thread panicked".into()))?
-        .unwrap_or_default();
-
-    let exit_code = status.code().unwrap_or(1);
-    if status.success() || !stderr_indicates_unknown_session(&stderr_bytes) {
-        return Ok(exit_code);
-    }
-
-    // Daemon told us our session id is gone. Allocate a fresh one and
-    // retry the wrapper invocation once.
-    let new_session_id = match allocate_replacement_session(zccache) {
-        Ok(id) => id,
-        Err(err) => {
-            eprintln!(
-                "soldr: zccache reported \"unknown session:\" but soldr could not allocate \
-                 a replacement session ({err}); propagating original exit code"
-            );
-            return Ok(exit_code);
-        }
-    };
-
-    eprintln!(
-        "soldr: zccache session resync after \"unknown session:\"; retrying once with fresh session {new_session_id}"
-    );
-
-    let mut retry = std::process::Command::new(zccache);
-    retry.args(&raw_args[1..]);
-    retry.env(soldr_cache::ZCCACHE_SESSION_ID_ENV_VAR, &new_session_id);
-    suppress_windows_console_window(&mut retry);
-    let retry_status = retry.status()?;
-    Ok(retry_status.code().unwrap_or(1))
-}
-
-/// Returns `true` iff `stderr` contains the literal substring
-/// `unknown session:` somewhere in its bytes. Tolerates non-UTF-8 input.
-///
-/// Extracted as a pure helper so the retry trigger can be unit-tested
-/// without spawning a real zccache.
-#[cfg_attr(unix, allow(dead_code))]
-fn stderr_indicates_unknown_session(stderr: &[u8]) -> bool {
-    const NEEDLE: &[u8] = b"unknown session:";
-    if stderr.len() < NEEDLE.len() {
-        return false;
-    }
-    stderr.windows(NEEDLE.len()).any(|w| w == NEEDLE)
-}
-
-/// Run `zccache session-start --stats --log <path> --journal <path>` against
-/// the cache dir the wrapper invocation inherits from cargo, and return the
-/// parsed session id. Mirrors the args used by `prepare_zccache_build`.
-///
-/// Used by the Windows wrapper retry path (issue #265): when the daemon
-/// reports `unknown session:` the in-process session id is stale, so soldr
-/// allocates a replacement before retrying the wrapper invocation once.
-#[cfg(not(unix))]
-fn allocate_replacement_session(zccache: &std::path::Path) -> Result<String, SoldrError> {
-    let cache_dir = std::env::var_os(soldr_cache::ZCCACHE_CACHE_DIR_ENV_VAR)
-        .map(std::path::PathBuf::from)
-        .ok_or_else(|| {
-            SoldrError::Other(format!(
-                "{} is not set in the wrapper environment; cannot allocate replacement zccache session",
-                soldr_cache::ZCCACHE_CACHE_DIR_ENV_VAR
-            ))
-        })?;
-
-    let session_log_path = soldr_cache::session_log_path(&cache_dir);
-    let session_log_path_arg = session_log_path.display().to_string();
-    let journal_path = soldr_cache::session_journal_path(&cache_dir);
-    let journal_path_arg = journal_path.display().to_string();
-    let session_json = run_zccache_command_in_cache_dir(
-        zccache,
-        &[
-            "session-start",
-            "--stats",
-            "--log",
-            &session_log_path_arg,
-            "--journal",
-            &journal_path_arg,
-        ],
-        &cache_dir,
-    )?;
-    soldr_cache::parse_zccache_session_id(&session_json.stdout).ok_or_else(|| {
-        SoldrError::Other(format!(
-            "failed to parse zccache session id from output: {}",
-            session_json.stdout.trim()
-        ))
-    })
-}
-
-/// Best-effort upsert of the workspace `target/` dir into the soldr
-/// state registry on every wrapper invocation. Silent on failure.
-///
-/// `rustc_args` is the slice of args that follows the rustc binary
-/// path in the wrapper invocation (i.e. `raw_args[2..]`).
-fn record_target_dir_in_registry(rustc_args: &[String]) {
-    let Some(target) = soldr_cache::target_registry::resolve_workspace_target_dir(rustc_args)
-    else {
-        return;
-    };
-    let Ok(paths) = SoldrPaths::new() else { return };
-    let db_path = soldr_cache::data_db_path(&paths);
-    let Ok(registry) = soldr_cache::target_registry::TargetRegistry::open(&db_path) else {
-        return;
-    };
-    let _ = registry.upsert(&target);
-}
-
-
-pub(crate) fn resolve_toolchain_binary(tool: &str) -> Result<std::path::PathBuf, SoldrError> {
-    if let Some(path) = toolchain_binary_override(tool) {
-        return Ok(path);
-    }
-
-    let start_dir = std::env::current_dir().ok();
-    if let Some(path) = probe_direct_toolchain_binary(tool, start_dir.as_deref()) {
-        return Ok(path);
-    }
-
-    let mut command = std::process::Command::new(rustup_binary());
-    command.args(["which", tool]);
-    apply_implicit_toolchain_homes(&mut command);
-    suppress_windows_console_window(&mut command);
-    let output = command.output();
-
-    match output {
-        Ok(output) if output.status.success() => {
-            let path = String::from_utf8_lossy(&output.stdout).trim().to_string();
-            if !path.is_empty() {
-                return Ok(path.into());
-            }
-        }
-        Ok(output) => {
-            if let Some(path) = soldr_core::probe_toolchain_binary(tool, start_dir.as_deref()) {
-                return Ok(path);
-            }
-            return Err(rustup_resolution_failure(tool, &output.stderr));
-        }
-        Err(err) => {
-            if let Some(path) = soldr_core::probe_toolchain_binary(tool, start_dir.as_deref()) {
-                return Ok(path);
-            }
-            return Err(SoldrError::Other(format!(
-                "failed to invoke rustup while resolving {tool}: {err}"
-            )));
-        }
-    }
-
-    if let Some(path) = soldr_core::probe_toolchain_binary(tool, start_dir.as_deref()) {
-        return Ok(path);
-    }
-
-    Err(SoldrError::Other(format!(
-        "rustup did not return a path for {tool}"
-    )))
-}
-
-fn probe_direct_toolchain_binary(
-    tool: &str,
-    start_dir: Option<&std::path::Path>,
-) -> Option<std::path::PathBuf> {
-    if std::env::var_os("RUSTUP_TOOLCHAIN").is_some_and(|value| !value.is_empty()) {
-        return None;
-    }
-
-    explicit_rustup_toolchain_binary(tool)
-        .or_else(|| repo_local_rustup_toolchain_binary(tool, start_dir))
-        .or_else(|| explicit_cargo_home_binary(tool))
-        .or_else(|| repo_local_cargo_home_binary(tool, start_dir))
-}
-
-fn explicit_cargo_home_binary(tool: &str) -> Option<std::path::PathBuf> {
-    non_empty_env_path("CARGO_HOME").and_then(|path| executable_in_dir(&path.join("bin"), tool))
-}
-
-fn repo_local_cargo_home_binary(
-    tool: &str,
-    start_dir: Option<&std::path::Path>,
-) -> Option<std::path::PathBuf> {
-    find_ancestor_dir(start_dir, ".cargo")
-        .and_then(|path| executable_in_dir(&path.join("bin"), tool))
-}
-
-fn explicit_rustup_toolchain_binary(tool: &str) -> Option<std::path::PathBuf> {
-    non_empty_env_path("RUSTUP_HOME")
-        .and_then(|path| rustup_home_single_toolchain_binary(&path, tool))
-}
-
-fn repo_local_rustup_toolchain_binary(
-    tool: &str,
-    start_dir: Option<&std::path::Path>,
-) -> Option<std::path::PathBuf> {
-    find_ancestor_dir(start_dir, ".rustup")
-        .and_then(|path| rustup_home_single_toolchain_binary(&path, tool))
-}
-
-fn rustup_home_single_toolchain_binary(
-    rustup_home: &std::path::Path,
-    tool: &str,
-) -> Option<std::path::PathBuf> {
-    let mut candidates = std::fs::read_dir(rustup_home.join("toolchains"))
-        .ok()?
-        .filter_map(Result::ok)
-        .map(|entry| entry.path().join("bin"))
-        .filter_map(|dir| executable_in_dir(&dir, tool))
-        .collect::<Vec<_>>();
-    if candidates.len() == 1 {
-        candidates.pop()
-    } else {
-        None
-    }
-}
-
-fn find_ancestor_dir(
-    start_dir: Option<&std::path::Path>,
-    relative: &str,
-) -> Option<std::path::PathBuf> {
-    let mut current = start_dir?.to_path_buf();
-    loop {
-        let candidate = current.join(relative);
-        if candidate.is_dir() {
-            return Some(candidate);
-        }
-        if !current.pop() {
-            return None;
-        }
-    }
-}
-
-fn executable_in_dir(dir: &std::path::Path, tool: &str) -> Option<std::path::PathBuf> {
-    let candidate = dir.join(tool);
-    if candidate.is_file() {
-        return Some(candidate);
-    }
-    #[cfg(windows)]
-    {
-        for suffix in windows_path_exts() {
-            let candidate = dir.join(format!("{tool}{suffix}"));
-            if candidate.is_file() {
-                return Some(candidate);
-            }
-        }
-    }
-    None
-}
-
-#[cfg(windows)]
-fn windows_path_exts() -> Vec<String> {
-    std::env::var_os("PATHEXT")
-        .and_then(|value| value.into_string().ok())
-        .unwrap_or_else(|| ".COM;.EXE;.BAT;.CMD".to_string())
-        .split(';')
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(|value| value.to_ascii_lowercase())
-        .collect()
-}
-
-pub(crate) fn apply_implicit_toolchain_homes(command: &mut std::process::Command) {
-    let start_dir = std::env::current_dir().ok();
-    soldr_core::apply_implicit_toolchain_homes(command, start_dir.as_deref());
-}
-
-fn rustup_resolution_failure(tool: &str, stderr: &[u8]) -> SoldrError {
-    let raw_failure = String::from_utf8_lossy(stderr).trim().to_string();
-    SoldrError::Other(format!(
-        "failed to resolve {tool} via rustup: {raw_failure}\n\
-CI hint: if this repository pins Rust in rust-toolchain.toml, preinstall that exact channel instead of a generic stable toolchain.\n\
-CI hint: export RUSTUP_TOOLCHAIN to that exact channel for later cargo, rustc, and soldr cargo steps, or use the documented setup-soldr action path (uses: zackees/soldr@<ref> or uses: ./)."
-    ))
-}
-
-fn parse_tool_spec(spec: &str) -> (String, VersionSpec) {
-    if let Some((name, version)) = spec.split_once('@') {
-        (name.to_string(), VersionSpec::parse(version))
-    } else {
-        (spec.to_string(), VersionSpec::Latest)
-    }
-}
-
-pub(crate) struct ZccacheBuildSession {
-    pub(crate) binary_path: std::path::PathBuf,
-    pub(crate) cache_dir: std::path::PathBuf,
-    pub(crate) session_id: String,
-    pub(crate) session_log_path: std::path::PathBuf,
-    pub(crate) journal_path: std::path::PathBuf,
-    pub(crate) session_stats_path: std::path::PathBuf,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum RustcWrapperMode {
-    ManagedZccache,
-    Custom(std::ffi::OsString),
-    Disabled,
-}
-
-fn rustc_wrapper_mode_from_env_var(value: Option<&std::ffi::OsStr>) -> RustcWrapperMode {
-    match value.and_then(std::ffi::OsStr::to_str) {
-        None => value
-            .map(|value| RustcWrapperMode::Custom(value.to_os_string()))
-            .unwrap_or(RustcWrapperMode::ManagedZccache),
-        Some(value) => {
-            let trimmed = value.trim();
-            if trimmed.is_empty() || trimmed.eq_ignore_ascii_case("none") {
-                RustcWrapperMode::Disabled
-            } else {
-                RustcWrapperMode::Custom(trimmed.into())
-            }
-        }
-    }
-}
-
-fn rustc_wrapper_mode() -> RustcWrapperMode {
-    rustc_wrapper_mode_from_env_var(std::env::var_os(RUSTC_WRAPPER_OVERRIDE_ENV_VAR).as_deref())
-}
-
-pub(crate) async fn prepare_rustc_wrapper(
-    cargo: &mut std::process::Command,
-    paths: &SoldrPaths,
-) -> Result<Option<ZccacheBuildSession>, SoldrError> {
-    match rustc_wrapper_mode() {
-        RustcWrapperMode::ManagedZccache => prepare_zccache_build(cargo, paths).await.map(Some),
-        RustcWrapperMode::Custom(wrapper) => {
-            if is_sccache_wrapper(&wrapper) && std::env::var_os("SCCACHE_DIR").is_none() {
-                let sccache_dir = soldr_cache::sccache_dir(paths);
-                std::fs::create_dir_all(&sccache_dir)?;
-                cargo.env("SCCACHE_DIR", sccache_dir);
-            }
-            cargo.env("RUSTC_WRAPPER", wrapper);
-            cargo.env_remove(soldr_cache::ZCCACHE_BINARY_ENV_VAR);
-            cargo.env_remove(soldr_cache::MANAGED_ZCCACHE_CACHE_DIR_ENV_VAR);
-            cargo.env_remove(soldr_cache::ZCCACHE_SESSION_ID_ENV_VAR);
-            Ok(None)
-        }
-        RustcWrapperMode::Disabled => {
-            cargo.env_remove("RUSTC_WRAPPER");
-            cargo.env_remove(soldr_cache::ZCCACHE_BINARY_ENV_VAR);
-            cargo.env_remove(soldr_cache::MANAGED_ZCCACHE_CACHE_DIR_ENV_VAR);
-            cargo.env_remove(soldr_cache::ZCCACHE_SESSION_ID_ENV_VAR);
-            Ok(None)
-        }
-    }
-}
-
-fn is_sccache_wrapper(wrapper: &std::ffi::OsStr) -> bool {
-    std::path::Path::new(wrapper)
-        .file_stem()
-        .and_then(std::ffi::OsStr::to_str)
-        .is_some_and(|stem| stem.eq_ignore_ascii_case("sccache"))
-}
-
-async fn prepare_zccache_build(
-    cargo: &mut std::process::Command,
-    paths: &SoldrPaths,
-) -> Result<ZccacheBuildSession, SoldrError> {
-    let zccache_dir = managed_zccache_cache_dir(paths)?;
-    std::fs::create_dir_all(&zccache_dir)?;
-    std::fs::create_dir_all(zccache_dir.join("logs"))?;
-    let fetch = fetch_managed_zccache(paths).await?;
-    if fetch.cached {
-        eprintln!(
-            "soldr: using managed zccache {}",
-            soldr_fetch::MANAGED_ZCCACHE_VERSION
-        );
-    } else {
-        eprintln!(
-            "soldr: fetched managed zccache {}",
-            soldr_fetch::MANAGED_ZCCACHE_VERSION
-        );
-    }
-
-    start_zccache_with_recovery(&fetch.binary_path, &zccache_dir)?;
-
-    let session_log_path = soldr_cache::session_log_path(&zccache_dir);
-    let session_log_path_arg = session_log_path.display().to_string();
-    let journal_path = soldr_cache::session_journal_path(&zccache_dir);
-    let journal_path_arg = journal_path.display().to_string();
-    let session_stats_path = soldr_cache::session_stats_path(&zccache_dir);
-    let session_json = run_zccache_command_in_cache_dir(
-        &fetch.binary_path,
-        &[
-            "session-start",
-            "--stats",
-            "--log",
-            &session_log_path_arg,
-            "--journal",
-            &journal_path_arg,
-        ],
-        &zccache_dir,
-    )?;
-    let session_id =
-        soldr_cache::parse_zccache_session_id(&session_json.stdout).ok_or_else(|| {
-            SoldrError::Other(format!(
-                "failed to parse zccache session id from output: {}",
-                session_json.stdout.trim()
-            ))
-        })?;
-
-    cargo.env("RUSTC_WRAPPER", current_soldr_binary()?);
-    cargo.env(soldr_cache::ZCCACHE_BINARY_ENV_VAR, &fetch.binary_path);
-    cargo.env(soldr_cache::ZCCACHE_CACHE_DIR_ENV_VAR, &zccache_dir);
-    cargo.env(soldr_cache::MANAGED_ZCCACHE_CACHE_DIR_ENV_VAR, &zccache_dir);
-    cargo.env(soldr_cache::ZCCACHE_SESSION_ID_ENV_VAR, &session_id);
-
-    Ok(ZccacheBuildSession {
-        binary_path: fetch.binary_path,
-        cache_dir: zccache_dir,
-        session_id,
-        session_log_path,
-        journal_path,
-        session_stats_path,
-    })
-}
-
-pub(crate) fn finish_zccache_build(session: &ZccacheBuildSession) -> Result<(), SoldrError> {
-    let output = run_zccache_command_raw_in_cache_dir(
-        &session.binary_path,
-        &["session-end", &session.session_id, "--json"],
-        &session.cache_dir,
-    )?;
-    if session.session_log_path.exists() {
-        eprintln!(
-            "soldr: zccache session log: {}",
-            session.session_log_path.display()
-        );
-    }
-    if output.status.success() {
-        let stdout = String::from_utf8_lossy(&output.stdout);
-        let stats_json = stdout.trim();
-        if !stats_json.is_empty() {
-            write_zccache_session_stats_json(session, stats_json)?;
-            let stats = parse_zccache_session_stats_json(stats_json)?;
-            print_zccache_session_stats(&stats, &session.session_stats_path);
-        }
-        return Ok(());
-    }
-
-    if zccache_json_flag_unsupported(&output) {
-        eprintln!(
-            "soldr: zccache JSON session summary unavailable; falling back to text session-end"
-        );
-        finish_zccache_build_text_fallback(session)?;
-        return Ok(());
-    }
-
-    Err(SoldrError::Other(zccache_command_failure_message(
-        &["session-end", &session.session_id, "--json"],
-        &output,
-    )))
-}
-
-fn finish_zccache_build_text_fallback(session: &ZccacheBuildSession) -> Result<(), SoldrError> {
-    let output = run_zccache_command_in_cache_dir(
-        &session.binary_path,
-        &["session-end", &session.session_id],
-        &session.cache_dir,
-    )?;
-    let stdout = output.stdout.trim();
-    if !stdout.is_empty() {
-        eprintln!("soldr: zccache session summary");
-        eprintln!("{stdout}");
-    }
-    Ok(())
-}
-
-fn write_zccache_session_stats_json(
-    session: &ZccacheBuildSession,
-    stats_json: &str,
-) -> Result<(), SoldrError> {
-    if let Some(parent) = session.session_stats_path.parent() {
-        std::fs::create_dir_all(parent)?;
-    }
-    std::fs::write(&session.session_stats_path, stats_json)?;
-    Ok(())
-}
-
-fn parse_zccache_session_stats_json(stats_json: &str) -> Result<serde_json::Value, SoldrError> {
-    serde_json::from_str(stats_json).map_err(|err| {
-        SoldrError::Other(format!(
-            "failed to parse zccache JSON session summary: {err}"
-        ))
-    })
-}
-
-fn print_zccache_session_stats(stats: &serde_json::Value, stats_path: &std::path::Path) {
-    eprintln!("soldr: zccache session summary");
-    eprintln!("  stats file: {}", stats_path.display());
-    match stats.get("status").and_then(serde_json::Value::as_str) {
-        Some("ok") => {
-            let hits = json_u64(stats, "hits").unwrap_or(0);
-            let misses = json_u64(stats, "misses").unwrap_or(0);
-            let non_cacheable = json_u64(stats, "non_cacheable").unwrap_or(0);
-            let errors = json_u64(stats, "errors").unwrap_or(0);
-            let compilations = json_u64(stats, "compilations").unwrap_or(hits + misses);
-            eprintln!(
-                "  compilations: {compilations}; hits: {hits}; misses: {misses}; non-cacheable: {non_cacheable}; errors: {errors}"
-            );
-            if let Some(hit_rate) = json_f64(stats, "hit_rate") {
-                eprintln!("  hit rate: {:.1}%", hit_rate * 100.0);
-            } else {
-                eprintln!("  hit rate: n/a");
-            }
-            let unique_sources = json_u64(stats, "unique_sources").unwrap_or(0);
-            let bytes_read = json_u64(stats, "bytes_read").unwrap_or(0);
-            let bytes_written = json_u64(stats, "bytes_written").unwrap_or(0);
-            eprintln!(
-                "  unique sources: {unique_sources}; bytes read: {bytes_read}; bytes written: {bytes_written}"
-            );
-            let time_saved_ms = json_u64(stats, "time_saved_ms").unwrap_or(0);
-            let duration_ms = json_u64(stats, "duration_ms").unwrap_or(0);
-            eprintln!("  time saved: {time_saved_ms} ms; duration: {duration_ms} ms");
-        }
-        Some("unavailable") => {
-            let reason = stats
-                .get("reason")
-                .and_then(serde_json::Value::as_str)
-                .unwrap_or("unknown");
-            eprintln!("  status: unavailable ({reason})");
-        }
-        Some("error") => {
-            let error = stats
-                .get("error")
-                .and_then(serde_json::Value::as_str)
-                .unwrap_or("unknown error");
-            eprintln!("  status: error ({error})");
-        }
-        Some(status) => eprintln!("  status: {status}"),
-        None => eprintln!("  status: unknown"),
-    }
-}
-
-fn json_u64(value: &serde_json::Value, key: &str) -> Option<u64> {
-    value.get(key).and_then(serde_json::Value::as_u64)
-}
-
-fn json_f64(value: &serde_json::Value, key: &str) -> Option<f64> {
-    value.get(key).and_then(serde_json::Value::as_f64)
-}
-
-fn zccache_json_flag_unsupported(output: &std::process::Output) -> bool {
-    let stderr = String::from_utf8_lossy(&output.stderr).to_ascii_lowercase();
-    stderr.contains("unexpected argument")
-        || stderr.contains("unrecognized option")
-        || stderr.contains("found argument")
-}
-
-
-pub(crate) struct CommandOutput {
-    pub(crate) stdout: String,
-}
-
-pub(crate) fn managed_zccache_cache_dir(paths: &SoldrPaths) -> Result<std::path::PathBuf, SoldrError> {
-    let zccache_dir = normalize_path_for_compare(&soldr_cache::zccache_dir(paths))?;
-    let inherited_soldr_managed_dir =
-        non_empty_env_path(soldr_cache::MANAGED_ZCCACHE_CACHE_DIR_ENV_VAR)
-            .map(|path| normalize_path_for_compare(&path))
-            .transpose()?;
-    if let Some(explicit) = non_empty_env_path(soldr_cache::ZCCACHE_CACHE_DIR_ENV_VAR) {
-        let explicit = normalize_path_for_compare(&explicit)?;
-        if explicit != zccache_dir && inherited_soldr_managed_dir.as_ref() != Some(&explicit) {
-            return Err(SoldrError::Other(format!(
-                "{} is managed by soldr for managed zccache builds. Unset it, set SOLDR_CACHE_DIR to choose soldr's cache root, or set SOLDR_RUSTC_WRAPPER to use a custom wrapper.",
-                soldr_cache::ZCCACHE_CACHE_DIR_ENV_VAR
-            )));
-        }
-    }
-    Ok(zccache_dir)
-}
-
-pub(crate) fn normalize_path_for_compare(path: &std::path::Path) -> Result<std::path::PathBuf, SoldrError> {
-    if path.is_absolute() {
-        Ok(path.to_path_buf())
-    } else {
-        Ok(std::env::current_dir()?.join(path))
-    }
-}
-
-pub(crate) fn run_zccache_command_in_cache_dir(
-    binary: &std::path::Path,
-    args: &[&str],
-    cache_dir: &std::path::Path,
-) -> Result<CommandOutput, SoldrError> {
-    run_zccache_command_with_env(
-        binary,
-        args,
-        &[(
-            soldr_cache::ZCCACHE_CACHE_DIR_ENV_VAR,
-            cache_dir.as_os_str(),
-        )],
-    )
-}
-
-pub(crate) fn run_zccache_command_strings_in_cache_dir(
-    binary: &std::path::Path,
-    args: &[String],
-    cache_dir: &std::path::Path,
-) -> Result<CommandOutput, SoldrError> {
-    let output = run_zccache_command_raw_strings_with_env(
-        binary,
-        args,
-        &[(
-            soldr_cache::ZCCACHE_CACHE_DIR_ENV_VAR,
-            cache_dir.as_os_str(),
-        )],
-    )?;
-    if !output.status.success() {
-        return Err(SoldrError::Other(zccache_command_failure_message_strings(
-            args, &output,
-        )));
-    }
-
-    Ok(CommandOutput {
-        stdout: String::from_utf8_lossy(&output.stdout).to_string(),
-    })
-}
-
-fn start_zccache_with_recovery(
-    binary: &std::path::Path,
-    cache_dir: &std::path::Path,
-) -> Result<(), SoldrError> {
-    let start = run_zccache_command_raw_in_cache_dir(binary, &["start"], cache_dir)?;
-    if start.status.success() {
-        return Ok(());
-    }
-
-    let initial_stderr = command_stderr(&start);
-    if !is_stale_zccache_daemon_start_failure(&initial_stderr) {
-        return Err(SoldrError::Other(zccache_command_failure_message(
-            &["start"],
-            &start,
-        )));
-    }
-
-    eprintln!(
-        "soldr: zccache start reported an unresponsive daemon; stopping stale state and retrying"
-    );
-    let stop_diagnostic = match run_zccache_command_raw_in_cache_dir(binary, &["stop"], cache_dir) {
-        Ok(stop) if stop.status.success() => None,
-        Ok(stop) => Some(zccache_command_failure_message(&["stop"], &stop)),
-        Err(err) => Some(format!("failed to invoke zccache stop: {err}")),
-    };
-
-    match run_zccache_command_raw_in_cache_dir(binary, &["start"], cache_dir) {
-        Ok(retry) if retry.status.success() => Ok(()),
-        Ok(retry) => {
-            let mut message = format!(
-                "zccache start failed after stale daemon recovery retry: {}",
-                command_stderr(&retry)
-            );
-            message.push_str(&format!(
-                "\ninitial zccache start failure: {}",
-                initial_stderr
-            ));
-            if let Some(stop_diagnostic) = stop_diagnostic {
-                message.push_str(&format!("\nzccache stop diagnostic: {stop_diagnostic}"));
-            }
-            Err(SoldrError::Other(message))
-        }
-        Err(err) => {
-            let mut message =
-                format!("failed to invoke zccache start during stale daemon recovery retry: {err}");
-            message.push_str(&format!(
-                "\ninitial zccache start failure: {}",
-                initial_stderr
-            ));
-            if let Some(stop_diagnostic) = stop_diagnostic {
-                message.push_str(&format!("\nzccache stop diagnostic: {stop_diagnostic}"));
-            }
-            Err(SoldrError::Other(message))
-        }
-    }
-}
-
-fn is_stale_zccache_daemon_start_failure(stderr: &str) -> bool {
-    let stderr = stderr.to_ascii_lowercase();
-    stderr.contains("not accepting connections")
-        || (stderr.contains("daemon process") && stderr.contains("exists"))
-}
-
-pub(crate) fn run_zccache_command_raw_in_cache_dir(
-    binary: &std::path::Path,
-    args: &[&str],
-    cache_dir: &std::path::Path,
-) -> Result<std::process::Output, SoldrError> {
-    run_zccache_command_raw_with_env(
-        binary,
-        args,
-        &[(
-            soldr_cache::ZCCACHE_CACHE_DIR_ENV_VAR,
-            cache_dir.as_os_str(),
-        )],
-    )
-}
-
-fn run_zccache_command_with_env(
-    binary: &std::path::Path,
-    args: &[&str],
-    envs: &[(&str, &std::ffi::OsStr)],
-) -> Result<CommandOutput, SoldrError> {
-    let output = run_zccache_command_raw_with_env(binary, args, envs)?;
-    if !output.status.success() {
-        return Err(SoldrError::Other(zccache_command_failure_message(
-            args, &output,
-        )));
-    }
-
-    Ok(CommandOutput {
-        stdout: String::from_utf8_lossy(&output.stdout).to_string(),
-    })
-}
-
-fn run_zccache_command_raw_with_env(
-    binary: &std::path::Path,
-    args: &[&str],
-    envs: &[(&str, &std::ffi::OsStr)],
-) -> Result<std::process::Output, SoldrError> {
-    let mut command = std::process::Command::new(binary);
-    command.args(args);
-    for &(name, value) in envs {
-        command.env(name, value);
-    }
-    suppress_windows_console_window(&mut command);
-    Ok(command.output()?)
-}
-
-fn run_zccache_command_raw_strings_with_env(
-    binary: &std::path::Path,
-    args: &[String],
-    envs: &[(&str, &std::ffi::OsStr)],
-) -> Result<std::process::Output, SoldrError> {
-    let mut command = std::process::Command::new(binary);
-    command.args(args);
-    for &(name, value) in envs {
-        command.env(name, value);
-    }
-    suppress_windows_console_window(&mut command);
-    Ok(command.output()?)
-}
-
-fn zccache_command_failure_message(args: &[&str], output: &std::process::Output) -> String {
-    format!(
-        "zccache {} failed: {}",
-        args.join(" "),
-        command_stderr(output)
-    )
-}
-
-fn zccache_command_failure_message_strings(
-    args: &[String],
-    output: &std::process::Output,
-) -> String {
-    format!(
-        "zccache {} failed: {}",
-        args.join(" "),
-        command_stderr(output)
-    )
-}
-
-pub(crate) fn command_stderr(output: &std::process::Output) -> String {
-    let stderr = String::from_utf8_lossy(&output.stderr).trim().to_owned();
-    if stderr.is_empty() {
-        format!("exit status {}", output.status)
-    } else {
-        stderr
-    }
-}
-
-fn toolchain_binary_override(tool: &str) -> Option<std::path::PathBuf> {
-    let env_var = match tool {
-        "cargo" => TEST_CARGO_BIN_ENV_VAR,
-        "rustc" => TEST_RUSTC_BIN_ENV_VAR,
-        _ => return real_toolchain_binary_override(tool),
-    };
-    non_empty_env_path(env_var).or_else(|| real_toolchain_binary_override(tool))
-}
-
-fn real_toolchain_binary_override(tool: &str) -> Option<std::path::PathBuf> {
-    non_empty_env_path(&real_toolchain_binary_env_var(tool))
-}
-
-fn real_toolchain_binary_env_var(tool: &str) -> String {
-    let mut value = String::from(REAL_TOOLCHAIN_BINARY_ENV_PREFIX);
-    for ch in tool.chars() {
-        if ch.is_ascii_alphanumeric() {
-            value.push(ch.to_ascii_uppercase());
-        } else {
-            value.push('_');
-        }
-    }
-    value
-}
-
-pub(crate) fn rustup_binary() -> std::path::PathBuf {
-    non_empty_env_path(TEST_RUSTUP_BIN_ENV_VAR).unwrap_or_else(|| "rustup".into())
-}
-
-fn zccache_binary_override() -> Option<std::path::PathBuf> {
-    non_empty_env_path(TEST_ZCCACHE_BIN_ENV_VAR)
-        .or_else(|| non_empty_env_path(soldr_cache::ZCCACHE_BINARY_ENV_VAR))
-}
-
-pub(crate) fn non_empty_env_path(env_var: &str) -> Option<std::path::PathBuf> {
-    let value = std::env::var_os(env_var)?;
-    if value.is_empty() {
-        return None;
-    }
-    Some(value.into())
-}
-
-fn current_soldr_binary() -> Result<std::path::PathBuf, SoldrError> {
-    std::env::current_exe().map_err(SoldrError::from)
-}
-
-async fn fetch_managed_zccache(paths: &SoldrPaths) -> Result<soldr_fetch::FetchResult, SoldrError> {
-    if let Some(binary_path) = non_empty_env_path(TEST_ZCCACHE_BIN_ENV_VAR) {
-        return Ok(soldr_fetch::FetchResult {
-            binary_path,
-            version: soldr_fetch::MANAGED_ZCCACHE_VERSION.to_string(),
-            cached: true,
-        });
-    }
-
-    soldr_fetch::fetch_zccache_with_paths(paths).await
-}
-
-pub(crate) fn cached_managed_zccache(
-    paths: &SoldrPaths,
-) -> Result<Option<soldr_fetch::FetchResult>, SoldrError> {
-    if let Some(binary_path) = non_empty_env_path(TEST_ZCCACHE_BIN_ENV_VAR) {
-        return Ok(Some(soldr_fetch::FetchResult {
-            binary_path,
-            version: soldr_fetch::MANAGED_ZCCACHE_VERSION.to_string(),
-            cached: true,
-        }));
-    }
-
-    soldr_fetch::cached_zccache_binary(paths)
 }
 
 #[cfg(test)]
