+++ crates/soldr-cli/src/tests.rs
@@ -1,7 +1,7 @@
 use super::cargo_front_door::{
-    cargo_args_specify_target, cargo_args_use_reserved_no_cache, cargo_profile, cargo_target_triple,
-    first_cargo_subcommand, low_disk_warning_for_free_bytes, low_disk_warning_for_path,
-    selected_cargo_args,
+    cargo_args_specify_target, cargo_args_use_reserved_no_cache, cargo_profile,
+    cargo_target_triple, first_cargo_subcommand, low_disk_warning_for_free_bytes,
+    low_disk_warning_for_path, selected_cargo_args,
 };
 use super::gc::{gc_purge_worker_count_for, parse_gc_purge_answer};
 use super::rust_plan::{
@@ -14,13 +14,15 @@ use super::rust_plan::{
     RustArtifactPlanContext, RustPlanInputs, RustPlanPackages, RustToolchainIdentity,
     ThinSliceManifest, WarmRestoreSentinel, WarmRestoreSkipInputs, BUNDLE_WALK_THREAD_CAP,
 };
+use super::wrapper::stderr_indicates_unknown_session;
+use super::zccache::{
+    is_sccache_wrapper, rustc_wrapper_mode_from_env_var, RustcWrapperMode, ZccacheBuildSession,
+};
 use super::{
-    extract_as_pin, is_sccache_wrapper, normalize_version, parse_tool_spec,
-    rustc_wrapper_mode_from_env_var, rustup_resolution_failure, should_self_relocate_for_invocation,
-    should_trampoline, stderr_indicates_unknown_session, Cli, Commands, GcSubcommand,
-    RustcWrapperMode, ZccacheBuildSession, LOW_DISK_WARNING_THRESHOLD_BYTES,
-    RUSTC_WRAPPER_OVERRIDE_ENV_VAR, SKIP_WARM_RESTORE_ENV_VAR, THIN_MANIFEST_FILENAME,
-    WARM_RESTORE_MAX_AGE_SECONDS,
+    extract_as_pin, normalize_version, parse_tool_spec, rustup_resolution_failure,
+    should_self_relocate_for_invocation, should_trampoline, Cli, Commands, GcSubcommand,
+    LOW_DISK_WARNING_THRESHOLD_BYTES, RUSTC_WRAPPER_OVERRIDE_ENV_VAR, SKIP_WARM_RESTORE_ENV_VAR,
+    THIN_MANIFEST_FILENAME, WARM_RESTORE_MAX_AGE_SECONDS,
 };
 use clap::Parser;
 use soldr_fetch::VersionSpec;
